reorganised components file structure
