+++ jada-app/src/Components/Charts/AppliedPercent/AppliedPercent.js
