+++ jada-app/src/Components/Charts/AppliedPercent/AppliedPercentConstructor/AppliedPercentConstructor.js
