+++ jada-app/src/Components/Charts/AppliedPercent/AppliedPercentConstructor/appliedPercentConstructor.css
