+++ jada-app/src/Components/Charts/AppliedPercent/appliedPercent.css
