+++ jada-app/src/Components/Charts/BarChartAvS/BarChartAvSConstructor/BarChartAvSConstructor.js
