+++ jada-app/src/Components/Charts/BarChartAvS/barChartAvS.css
