+++ jada-app/src/Components/Charts/BarChartDKW/barChartDKW.css
