+++ jada-app/src/Components/Charts/BarChartLocations/barChartLocations.css
