+++ jada-app/src/Components/Charts/BarChartTop24/BarChartTop24.js
@@ -1,6 +1,6 @@
 import React from "react";
 import './barChartTop24.css';
-import BarChartConstructor from "../BarChartConstructor/BarChartConstructor";
+import BarChartConstructor from "../../Constructors/BarChartConstructor/BarChartConstructor";
 
 class BarChartTop24 extends React.Component {
     constructor(props) {
