+++ jada-app/src/Components/Charts/BarChartUDKW/barChartUDKW.css
