+++ jada-app/src/Components/Constructors/BarChartConstructor/BarChartConstructor.js
