+++ jada-app/src/Components/Constructors/BarChartHorizontalContructor/BarChartHorizontalConstructor.js
