+++ jada-app/src/Components/Constructors/PercentBoxConstructor/percentBoxConstructor.css
